refactor: extract terminal logic and styles
