+++ src/components/AdminDashboard.css
@@ -1,0 +1,30 @@
+.admin-modal {
+  position: fixed;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  background: rgba(0, 0, 0, 0.9);
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  z-index: 1200;
+  backdrop-filter: blur(10px);
+}
+
+.admin-content {
+  background: #1a1a1a;
+  border: 2px solid #00d4aa;
+  border-radius: 10px;
+  padding: 20px;
+  max-width: 600px;
+  width: 90%;
+  font-family: 'Inter', sans-serif;
+  color: #eee;
+}
+
+.admin-content h2 {
+  margin: 0 0 10px;
+  color: #00d4aa;
+}
+
