+++ src/components/PaymentProcessor.css
@@ -1,0 +1,415 @@
+.payment-modal {
+  position: fixed;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  background: rgba(0, 0, 0, 0.95);
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  z-index: 1100;
+  backdrop-filter: blur(15px);
+  animation: modalFadeIn 0.3s ease;
+}
+
+.payment-content {
+  background: linear-gradient(135deg, #1a1a1a 0%, #2a2a2a 100%);
+  border: 2px solid #00d4aa;
+  border-radius: 12px;
+  padding: 30px;
+  max-width: 1000px;
+  width: 95%;
+  max-height: 90vh;
+  overflow-y: auto;
+  box-shadow: 0 20px 60px rgba(0, 212, 170, 0.4);
+  font-family: 'Inter', sans-serif;
+  animation: modalSlideIn 0.4s ease;
+}
+
+.payment-intro {
+  text-align: center;
+  margin-bottom: 30px;
+}
+
+.payment-intro p {
+  color: #888;
+  font-size: 16px;
+  margin-bottom: 15px;
+}
+
+.security-badges {
+  display: flex;
+  justify-content: center;
+  gap: 15px;
+  flex-wrap: wrap;
+}
+
+.badge {
+  background: rgba(0, 212, 170, 0.1);
+  border: 1px solid #00d4aa;
+  color: #00d4aa;
+  padding: 6px 12px;
+  border-radius: 20px;
+  font-size: 12px;
+  font-weight: 600;
+}
+
+.packages-grid {
+  display: grid;
+  grid-template-columns: repeat(auto-fit, minmax(280px, 1fr));
+  gap: 20px;
+  margin-bottom: 30px;
+}
+
+.package-card {
+  background: rgba(0, 0, 0, 0.5);
+  border: 2px solid #333;
+  border-radius: 12px;
+  padding: 25px;
+  cursor: pointer;
+  transition: all 0.3s ease;
+  position: relative;
+  overflow: hidden;
+}
+
+.package-card:hover {
+  border-color: #00d4aa;
+  box-shadow: 0 10px 30px rgba(0, 212, 170, 0.2);
+  transform: translateY(-3px);
+}
+
+.package-card:nth-child(2) {
+  border-color: #ffd93d;
+  background: rgba(255, 217, 61, 0.05);
+}
+
+.package-card:nth-child(2):hover {
+  border-color: #ffd93d;
+  box-shadow: 0 10px 30px rgba(255, 217, 61, 0.2);
+}
+
+.package-header {
+  margin-bottom: 15px;
+}
+
+.package-header h3 {
+  color: #00d4aa;
+  font-size: 18px;
+  margin-bottom: 8px;
+  font-weight: 700;
+}
+
+.package-price {
+  display: flex;
+  flex-direction: column;
+  gap: 4px;
+}
+
+.total-price {
+  color: #64ffda;
+  font-size: 24px;
+  font-weight: 700;
+}
+
+.downpayment {
+  color: #ffd93d;
+  font-size: 14px;
+  font-weight: 600;
+}
+
+.package-description {
+  color: #888;
+  margin-bottom: 15px;
+  line-height: 1.5;
+}
+
+.package-features {
+  list-style: none;
+  margin-bottom: 20px;
+}
+
+.package-features li {
+  color: #00d4aa;
+  padding: 4px 0;
+  padding-left: 20px;
+  position: relative;
+  font-size: 14px;
+}
+
+.package-features li:before {
+  content: '✓';
+  position: absolute;
+  left: 0;
+  color: #51cf66;
+  font-weight: bold;
+}
+
+.select-package-btn {
+  width: 100%;
+  background: linear-gradient(90deg, #00d4aa, #64ffda);
+  border: none;
+  border-radius: 8px;
+  padding: 12px 20px;
+  color: #000;
+  font-weight: 700;
+  cursor: pointer;
+  transition: all 0.3s ease;
+  text-transform: uppercase;
+  letter-spacing: 1px;
+  font-size: 14px;
+}
+
+.select-package-btn:hover {
+  transform: translateY(-2px);
+  box-shadow: 0 8px 20px rgba(0, 212, 170, 0.3);
+}
+
+.payment-terms {
+  color: #888;
+  font-size: 12px;
+  text-align: center;
+  line-height: 1.5;
+  border-top: 2px solid #333;
+  padding-top: 20px;
+}
+
+/* Payment Form Styles */
+.payment-summary {
+  background: rgba(0, 0, 0, 0.3);
+  border-radius: 8px;
+  padding: 20px;
+  margin-bottom: 30px;
+}
+
+.payment-summary h3 {
+  color: #64ffda;
+  margin-bottom: 15px;
+  font-size: 18px;
+}
+
+.summary-item {
+  display: flex;
+  justify-content: space-between;
+  margin-bottom: 10px;
+  padding: 8px 0;
+  border-bottom: 1px solid #333;
+}
+
+.summary-item:last-child {
+  border-bottom: none;
+}
+
+.summary-item span:first-child {
+  color: #888;
+}
+
+.summary-item span:last-child {
+  color: #00d4aa;
+  font-weight: 600;
+}
+
+.summary-item.downpayment span:last-child {
+  color: #ffd93d;
+  font-weight: 700;
+  font-size: 18px;
+}
+
+.summary-item.remaining span:last-child {
+  color: #888;
+  font-style: italic;
+}
+
+.payment-form-grid {
+  display: grid;
+  grid-template-columns: 1fr 1fr;
+  gap: 20px;
+}
+
+.payment-form-grid h3 {
+  grid-column: 1 / -1;
+  color: #64ffda;
+  margin-bottom: 15px;
+  font-size: 18px;
+  border-bottom: 2px solid #333;
+  padding-bottom: 10px;
+}
+
+.payment-security {
+  grid-column: 1 / -1;
+  margin-top: 20px;
+  text-align: center;
+}
+
+.security-info {
+  color: #888;
+  font-size: 14px;
+  margin-bottom: 20px;
+  padding: 15px;
+  background: rgba(0, 212, 170, 0.05);
+  border-radius: 8px;
+  border: 1px solid rgba(0, 212, 170, 0.2);
+}
+
+.process-payment-btn {
+  background: linear-gradient(90deg, #51cf66, #69db7c);
+  border: none;
+  border-radius: 8px;
+  padding: 18px 40px;
+  color: #000;
+  font-size: 18px;
+  font-weight: 700;
+  cursor: pointer;
+  transition: all 0.3s ease;
+  text-transform: uppercase;
+  letter-spacing: 1px;
+  box-shadow: 0 5px 15px rgba(81, 207, 102, 0.3);
+}
+
+.process-payment-btn:hover {
+  transform: translateY(-3px);
+  box-shadow: 0 10px 25px rgba(81, 207, 102, 0.4);
+  background: linear-gradient(90deg, #69db7c, #51cf66);
+}
+
+/* Processing Animation */
+.processing {
+  text-align: center;
+  padding: 60px 40px;
+}
+
+.processing-animation h2 {
+  color: #00d4aa;
+  margin-bottom: 30px;
+  font-size: 28px;
+}
+
+.processing-animation p {
+  color: #888;
+  margin-bottom: 10px;
+  font-size: 16px;
+}
+
+.spinner {
+  width: 80px;
+  height: 80px;
+  border: 4px solid #333;
+  border-top: 4px solid #00d4aa;
+  border-radius: 50%;
+  animation: spin 1s linear infinite;
+  margin: 0 auto 30px;
+}
+
+@keyframes spin {
+  0% { transform: rotate(0deg); }
+  100% { transform: rotate(360deg); }
+}
+
+/* Success Payment */
+.success-payment {
+  text-align: center;
+  padding: 40px;
+}
+
+.success-content {
+  max-width: 500px;
+  margin: 0 auto;
+}
+
+.success-icon {
+  font-size: 80px;
+  margin-bottom: 20px;
+  animation: successPulse 1.5s ease infinite;
+}
+
+.success-content h2 {
+  color: #51cf66;
+  font-size: 28px;
+  margin-bottom: 25px;
+  text-shadow: 0 0 10px rgba(81, 207, 102, 0.5);
+}
+
+.success-details {
+  background: rgba(0, 0, 0, 0.3);
+  border-radius: 8px;
+  padding: 20px;
+  margin-bottom: 25px;
+}
+
+.success-details p {
+  color: #888;
+  margin-bottom: 8px;
+  font-size: 16px;
+}
+
+.success-details strong {
+  color: #00d4aa;
+}
+
+.next-steps {
+  text-align: left;
+  background: rgba(0, 212, 170, 0.05);
+  border-radius: 8px;
+  padding: 20px;
+  margin-bottom: 25px;
+  border: 1px solid rgba(0, 212, 170, 0.2);
+}
+
+.next-steps h3 {
+  color: #64ffda;
+  margin-bottom: 15px;
+  font-size: 18px;
+}
+
+.next-steps ul {
+  list-style: none;
+}
+
+.next-steps li {
+  color: #888;
+  margin-bottom: 8px;
+  padding-left: 25px;
+  position: relative;
+  line-height: 1.5;
+}
+
+.next-steps li:before {
+  content: attr(data-icon);
+  position: absolute;
+  left: 0;
+  color: #00d4aa;
+}
+
+.close-success-btn {
+  background: linear-gradient(90deg, #00d4aa, #64ffda);
+  border: none;
+  border-radius: 8px;
+  padding: 15px 30px;
+  color: #000;
+  font-size: 16px;
+  font-weight: 700;
+  cursor: pointer;
+  transition: all 0.3s ease;
+  text-transform: uppercase;
+  letter-spacing: 1px;
+}
+
+.close-success-btn:hover {
+  transform: translateY(-2px);
+  box-shadow: 0 10px 20px rgba(0, 212, 170, 0.3);
+}
+
+@keyframes successPulse {
+  0%, 100% { transform: scale(1); }
+  50% { transform: scale(1.05); }
+}
+
+/* Mobile Payment Responsiveness */
+@media (max-width: 768px) {
+  .packages-grid {
+    grid-template-columns: 1fr;
+  }
+  
+  .payment-form-grid {
+    grid-template-columns: 1fr;
+  }
