+++ src/components/PaymentProcessor.jsx
@@ -1,4 +1,5 @@
 import React, { useState } from 'react'
+import './PaymentProcessor.css'
 import { loadStripe } from '@stripe/stripe-js'
 
 // Initialize Stripe
