+++ src/components/ROICalculator.css
@@ -1,0 +1,199 @@
+.roi-calculator-modal {
+  position: fixed;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  background: rgba(0, 0, 0, 0.9);
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  z-index: 1000;
+  backdrop-filter: blur(10px);
+  animation: modalFadeIn 0.3s ease;
+}
+
+.roi-calculator-content {
+  background: linear-gradient(135deg, #1a1a1a 0%, #2a2a2a 100%);
+  border: 2px solid #00d4aa;
+  border-radius: 12px;
+  padding: 30px;
+  max-width: 900px;
+  width: 90%;
+  max-height: 90vh;
+  overflow-y: auto;
+  box-shadow: 0 20px 40px rgba(0, 212, 170, 0.3);
+  font-family: 'Inter', sans-serif;
+  animation: modalSlideIn 0.4s ease;
+}
+
+.calculator-intro {
+  color: #888;
+  font-size: 16px;
+  margin-bottom: 25px;
+  text-align: center;
+}
+
+.calculate-btn {
+  width: 100%;
+  background: linear-gradient(90deg, #ffd93d, #ffb347);
+  border: none;
+  border-radius: 8px;
+  padding: 15px 30px;
+  color: #000;
+  font-family: 'Inter', sans-serif;
+  font-size: 16px;
+  font-weight: 700;
+  cursor: pointer;
+  transition: all 0.3s ease;
+  text-transform: uppercase;
+  letter-spacing: 1px;
+  margin-top: 20px;
+}
+
+.calculate-btn:hover:not(:disabled) {
+  transform: translateY(-2px);
+  box-shadow: 0 10px 20px rgba(255, 217, 61, 0.3);
+  background: linear-gradient(90deg, #ffb347, #ffd93d);
+}
+
+.calculate-btn:disabled {
+  opacity: 0.6;
+  cursor: not-allowed;
+}
+
+.roi-results {
+  text-align: center;
+}
+
+.roi-results h3 {
+  color: #00d4aa;
+  font-size: 24px;
+  margin-bottom: 30px;
+  text-shadow: 0 0 5px rgba(0, 212, 170, 0.5);
+}
+
+.results-grid {
+  display: grid;
+  grid-template-columns: repeat(auto-fit, minmax(200px, 1fr));
+  gap: 20px;
+  margin-bottom: 30px;
+}
+
+.result-card {
+  background: rgba(0, 0, 0, 0.5);
+  border: 2px solid #333;
+  border-radius: 12px;
+  padding: 20px;
+  transition: all 0.3s ease;
+}
+
+.result-card:hover {
+  border-color: #00d4aa;
+  box-shadow: 0 5px 15px rgba(0, 212, 170, 0.2);
+  transform: translateY(-2px);
+}
+
+.result-card.primary {
+  border-color: #ffd93d;
+  background: rgba(255, 217, 61, 0.1);
+}
+
+.result-card h4 {
+  color: #64ffda;
+  font-size: 14px;
+  margin-bottom: 10px;
+  text-transform: uppercase;
+  letter-spacing: 1px;
+}
+
+.result-value {
+  color: #00d4aa;
+  font-size: 28px;
+  font-weight: 700;
+  margin-bottom: 5px;
+  text-shadow: 0 0 5px rgba(0, 212, 170, 0.5);
+}
+
+.result-card.primary .result-value {
+  color: #ffd93d;
+  text-shadow: 0 0 5px rgba(255, 217, 61, 0.5);
+}
+
+.result-card p {
+  color: #888;
+  font-size: 12px;
+}
+
+.breakdown {
+  background: rgba(0, 0, 0, 0.3);
+  border-radius: 8px;
+  padding: 20px;
+  margin-bottom: 20px;
+}
+
+.breakdown h4 {
+  color: #64ffda;
+  margin-bottom: 15px;
+  font-size: 16px;
+}
+
+.breakdown-item {
+  display: flex;
+  justify-content: space-between;
+  margin-bottom: 10px;
+  padding: 8px 0;
+  border-bottom: 1px solid #333;
+}
+
+.breakdown-item:last-child {
+  border-bottom: none;
+}
+
+.breakdown-item span:first-child {
+  color: #888;
+}
+
+.breakdown-item span:last-child {
+  color: #00d4aa;
+  font-weight: 600;
+}
+
+.result-footer {
+  border-top: 2px solid #333;
+  padding-top: 20px;
+}
+
+.contact-btn {
+  width: 100%;
+  background: linear-gradient(90deg, #00d4aa, #64ffda);
+  border: none;
+  border-radius: 8px;
+  padding: 15px 30px;
+  color: #000;
+  font-family: 'Inter', sans-serif;
+  font-size: 16px;
+  font-weight: 700;
+  cursor: pointer;
+  transition: all 0.3s ease;
+  text-transform: uppercase;
+  letter-spacing: 1px;
+  margin-top: 10px;
+}
+
+.contact-btn:hover {
+  transform: translateY(-2px);
+  box-shadow: 0 10px 20px rgba(0, 212, 170, 0.3);
+  background: linear-gradient(90deg, #64ffda, #00d4aa);
+}
+
+@media (max-width: 768px) {
+  .roi-calculator-content {
+    padding: 20px;
+    margin: 20px;
+    max-width: calc(100% - 40px);
+  }
+  .results-grid {
+    grid-template-columns: 1fr;
+  }
+}
